+++ lib/src/multi_pull.dart
@@ -165,7 +165,8 @@ class MultiPull extends StatefulWidget {
 
 /// Contains the state for a [RefreshIndicator]. This class can be used to
 /// programmatically show the refresh indicator, see the [show] method.
-class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<MultiPull> {
+class MultiPullState extends State<MultiPull>
+    with TickerProviderStateMixin<MultiPull> {
   late AnimationController _positionController;
   late AnimationController _horizonPositionController;
   late AnimationController _scaleController;
@@ -188,9 +189,12 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
 
   late Widget _indicator;
 
-  static final Animatable<double> _threeQuarterTween = Tween<double>(begin: 0.0, end: 0.75);
-  static final Animatable<double> _kDragSizeFactorLimitTween = Tween<double>(begin: 0.0, end: _kDragSizeFactorLimit);
-  static final Animatable<double> _oneToZeroTween = Tween<double>(begin: 1.0, end: 0.0);
+  static final Animatable<double> _threeQuarterTween =
+      Tween<double>(begin: 0.0, end: 0.75);
+  static final Animatable<double> _kDragSizeFactorLimitTween =
+      Tween<double>(begin: 0.0, end: _kDragSizeFactorLimit);
+  static final Animatable<double> _oneToZeroTween =
+      Tween<double>(begin: 1.0, end: 0.0);
 
   @override
   void initState() {
@@ -231,7 +235,10 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
 
   bool _handleScrollNotification(ScrollNotification notification) {
     if (!widget.notificationPredicate(notification)) return false;
-    if (notification is ScrollStartNotification && notification.metrics.extentBefore == 0.0 && _mode == null && _start(notification.metrics.axisDirection)) {
+    if (notification is ScrollStartNotification &&
+        notification.metrics.extentBefore == 0.0 &&
+        _mode == null &&
+        _start(notification.metrics.axisDirection)) {
       setState(() {
         _mode = _RefreshIndicatorMode.drag;
 
@@ -259,9 +266,12 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
         break;
     }
     if (indicatorAtTopNow != _isIndicatorAtTop) {
-      if (_mode == _RefreshIndicatorMode.drag || _mode == _RefreshIndicatorMode.armed) _dismiss(_RefreshIndicatorMode.canceled);
+      if (_mode == _RefreshIndicatorMode.drag ||
+          _mode == _RefreshIndicatorMode.armed)
+        _dismiss(_RefreshIndicatorMode.canceled);
     } else if (notification is ScrollUpdateNotification) {
-      if (_mode == _RefreshIndicatorMode.drag || _mode == _RefreshIndicatorMode.armed) {
+      if (_mode == _RefreshIndicatorMode.drag ||
+          _mode == _RefreshIndicatorMode.armed) {
         if (notification.metrics.extentBefore > 0.0) {
           _dismiss(_RefreshIndicatorMode.canceled);
         } else {
@@ -274,11 +284,13 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
           _checkDragOffset(notification.dragDetails);
         }
       }
-      if (_mode == _RefreshIndicatorMode.armed && notification.dragDetails == null) {
+      if (_mode == _RefreshIndicatorMode.armed &&
+          notification.dragDetails == null) {
         _show();
       }
     } else if (notification is OverscrollNotification) {
-      if (_mode == _RefreshIndicatorMode.drag || _mode == _RefreshIndicatorMode.armed) {
+      if (_mode == _RefreshIndicatorMode.drag ||
+          _mode == _RefreshIndicatorMode.armed) {
         if (_dragOffset != null) {
           _dragOffset = _dragOffset! - notification.overscroll / 2.0;
         }
@@ -339,10 +351,13 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
   }
 
   void _checkDragOffset(DragUpdateDetails? details) {
-    assert(_mode == _RefreshIndicatorMode.drag || _mode == _RefreshIndicatorMode.armed);
+    assert(_mode == _RefreshIndicatorMode.drag ||
+        _mode == _RefreshIndicatorMode.armed);
     if (details == null) return;
-    double newValue = _dragOffset! / (details.globalPosition.dy * _kDragContainerExtentPercentage);
-    if (_mode == _RefreshIndicatorMode.armed) newValue = math.max(newValue, 1.0 / _kDragSizeFactorLimit);
+    double newValue = _dragOffset! /
+        (details.globalPosition.dy * _kDragContainerExtentPercentage);
+    if (_mode == _RefreshIndicatorMode.armed)
+      newValue = math.max(newValue, 1.0 / _kDragSizeFactorLimit);
     _positionController.value = newValue.clamp(0.0, 1.0);
 
     if (_mode == _RefreshIndicatorMode.armed) {
@@ -359,7 +374,8 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
       }
     }
 
-    if (_mode == _RefreshIndicatorMode.drag && _valueColor.value!.alpha == 0xFF) {
+    if (_mode == _RefreshIndicatorMode.drag &&
+        _valueColor.value!.alpha == 0xFF) {
       _mode = _RefreshIndicatorMode.armed;
     }
   }
@@ -367,17 +383,21 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
   // Stop showing the refresh indicator.
   Future<void> _dismiss(_RefreshIndicatorMode newMode) async {
     await Future<void>.value();
-    assert(newMode == _RefreshIndicatorMode.canceled || newMode == _RefreshIndicatorMode.done);
+    assert(newMode == _RefreshIndicatorMode.canceled ||
+        newMode == _RefreshIndicatorMode.done);
     setState(() {
       _mode = newMode;
     });
     switch (_mode) {
       case _RefreshIndicatorMode.done:
-        await _scaleController.animateTo(1.0, duration: _kIndicatorScaleDuration);
+        await _scaleController.animateTo(1.0,
+            duration: _kIndicatorScaleDuration);
         break;
       case _RefreshIndicatorMode.canceled:
-        await _positionController.animateTo(0.0, duration: _kIndicatorScaleDuration);
-        await _horizonPositionController.animateTo(0.0, duration: _kIndicatorScaleDuration);
+        await _positionController.animateTo(0.0,
+            duration: _kIndicatorScaleDuration);
+        await _horizonPositionController.animateTo(0.0,
+            duration: _kIndicatorScaleDuration);
         break;
       default:
         assert(false);
@@ -416,10 +436,13 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
           _mode = _RefreshIndicatorMode.refresh;
         });
       }
-      final bool showIndeterminateIndicator = _mode == _RefreshIndicatorMode.refresh || _mode == _RefreshIndicatorMode.done;
+      final bool showIndeterminateIndicator =
+          _mode == _RefreshIndicatorMode.refresh ||
+              _mode == _RefreshIndicatorMode.done;
       setState(() {
         _indicator = RefreshProgressIndicator(
-          semanticsLabel: widget.semanticsLabel ?? MaterialLocalizations.of(context).refreshIndicatorSemanticLabel,
+          semanticsLabel: widget.semanticsLabel ??
+              MaterialLocalizations.of(context).refreshIndicatorSemanticLabel,
           semanticsValue: widget.semanticsValue,
           value: showIndeterminateIndicator ? null : _value.value,
           valueColor: _valueColor,
@@ -464,7 +487,8 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
   /// actual scroll view. It defaults to showing the indicator at the top. To
   /// show it at the bottom, set `atTop` to false.
   Future<void> show({bool atTop = true}) {
-    if (_mode != _RefreshIndicatorMode.refresh && _mode != _RefreshIndicatorMode.snap) {
+    if (_mode != _RefreshIndicatorMode.refresh &&
+        _mode != _RefreshIndicatorMode.snap) {
       if (_mode == null) _start(atTop ? AxisDirection.down : AxisDirection.up);
       _show();
     }
@@ -508,8 +532,12 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
               axisAlignment: _isIndicatorAtTop! ? 1.0 : -1.0,
               sizeFactor: _positionFactor, // this is what brings it down
               child: Container(
-                padding: _isIndicatorAtTop! ? EdgeInsets.only(top: widget.displacement) : EdgeInsets.only(bottom: widget.displacement),
-                alignment: _isIndicatorAtTop! ? Alignment.topCenter : Alignment.bottomCenter,
+                padding: _isIndicatorAtTop!
+                    ? EdgeInsets.only(top: widget.displacement)
+                    : EdgeInsets.only(bottom: widget.displacement),
+                alignment: _isIndicatorAtTop!
+                    ? Alignment.topCenter
+                    : Alignment.bottomCenter,
                 child: ScaleTransition(
                   scale: _scaleFactor,
                   child: AnimatedBuilder(
@@ -540,8 +568,12 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
               axisAlignment: _isIndicatorAtTop! ? 1.0 : -1.0,
               sizeFactor: _positionFactor, // this is what brings it down
               child: Container(
-                padding: _isIndicatorAtTop! ? EdgeInsets.only(top: widget.displacement) : EdgeInsets.only(bottom: widget.displacement),
-                alignment: _isIndicatorAtTop! ? Alignment.topCenter : Alignment.bottomCenter,
+                padding: _isIndicatorAtTop!
+                    ? EdgeInsets.only(top: widget.displacement)
+                    : EdgeInsets.only(bottom: widget.displacement),
+                alignment: _isIndicatorAtTop!
+                    ? Alignment.topCenter
+                    : Alignment.bottomCenter,
                 child: ScaleTransition(
                   scale: _scaleFactor,
                   child: AnimatedBuilder(
@@ -549,7 +581,10 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
                       builder: (context, child) {
                         return Transform.translate(
                           child: Opacity(
-                            opacity: _mode == _RefreshIndicatorMode.refresh || _mode == _RefreshIndicatorMode.done ? 0.0 : widget.circleOpacity,
+                            opacity: _mode == _RefreshIndicatorMode.refresh ||
+                                    _mode == _RefreshIndicatorMode.done
+                                ? 0.0
+                                : widget.circleOpacity,
                             child: Container(
                               width: _actionSize,
                               height: _actionSize,
@@ -560,7 +595,9 @@ class MultiPullState extends State<MultiPull> with TickerProviderStateMixin<Mult
                             ),
                           ),
                           offset: Offset(
-                            (_horizonPositionController.value * indicatorWidth) - (indicatorWidth / 2),
+                            (_horizonPositionController.value *
+                                    indicatorWidth) -
+                                (indicatorWidth / 2),
                             0,
                           ),
                         );